fix: apply review feedback — uint32_t durationMs, override validation, motor TTL auto-stop
Agent-Logs-Url: https://github.com/tamaygz/esp-rollaball-derby/sessions/0abca415-9784-46e1-ab76-69504b15a94a

+++ clients/esp32-motor/src/websocket.cpp
@@ -244,7 +244,7 @@ void WSClient::_onMessage(WebsocketsMessage msg) {
         }
         msg.speedMs    = (uint16_t)(doc["payload"]["params"]["speed"]      | 1000);
         msg.brightness = (uint8_t) (doc["payload"]["params"]["brightness"] | 255);
-        msg.durationMs = (uint16_t)(doc["payload"]["durationMs"]           | 0);
+        msg.durationMs = (uint32_t)(doc["payload"]["durationMs"]           | 0);
         const char* text = doc["payload"]["params"]["text"] | "";
         strlcpy(msg.text, text, sizeof(msg.text));
         _pendingTestEffect    = msg;
